rename(package):fixlog로 패키지 이름 변경
