--- src/main/java/com/example/fixLog/controller/MemberController.java
@@ -1,4 +1,0 @@
-package com.example.fixLog.controller;
-
-public class MemberController {
-}
