--- src/main/java/com/example/fixLog/service/MemberService.java
@@ -1,4 +1,0 @@
-package com.example.fixLog.service;
-
-public class MemberService {
-}
