+++ src/main/java/com/example/fixlog/controller/MemberController.java
@@ -1,0 +1,4 @@
+package com.example.fixlog.controller;
+
+public class MemberController {
+}
