+++ src/main/java/com/example/fixlog/domain/post/Post.java
@@ -1,6 +1,6 @@
-package com.example.fixLog.domain.post;
+package com.example.fixlog.domain.post;
 
-import com.example.fixLog.domain.member.Member;
+import com.example.fixlog.domain.member.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
