+++ src/main/java/com/example/fixlog/dto/follow/request/UnfollowRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.fixLog.dto.follow.request;
+package com.example.fixlog.dto.follow.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class FollowRequestDto {
+public class UnfollowRequestDto {
     @JsonProperty("target_member_id")
-    private Long targetMemberId;  // 팔로우 대상 ID
+    private Long targetMemberId;
 }
