+++ src/main/java/com/example/fixlog/dto/follow/response/FollowResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.fixLog.dto.follow.response;
+package com.example.fixlog.dto.follow.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
@@ -6,7 +6,7 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class FollowingListResponseDto {
+public class FollowResponseDto {
     @JsonProperty("follow_id")
     private Long followId;
