+++ src/main/java/com/example/fixlog/dto/follow/response/FollowerListResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.fixLog.dto.follow.response;
+package com.example.fixlog.dto.follow.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
@@ -6,12 +6,12 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class FollowResponseDto {
+public class FollowerListResponseDto {
     @JsonProperty("follow_id")
     private Long followId;
 
-    @JsonProperty("following_id")
-    private Long followingId;
+    @JsonProperty("follower_id")
+    private Long followerId;
 
     private String nickname;
 }
