+++ src/main/java/com/example/fixlog/service/MemberService.java
@@ -1,0 +1,4 @@
+package com.example.fixlog.service;
+
+public class MemberService {
+}
