+++ src/test/java/com/example/fixlog/FixLogApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.fixLog;
+package com.example.fixlog;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
